lots of changes
there's an invisible div containing text using each of the fonts, so
they're preloaded. Can't be bothered to work out how to use the
FontFaceSet API.

Clicking the finish button downloads the SVG. Still need to make it get
the filename of the input image.

You can drag and drop images onto the page.

The mask is drawn at opacity 0.1.

Fixed the padding on the controls so the page doesn't scroll.